Reorganize the authorize page. Bunch of design fixes for the claim page.
