+++ apps/studio/components/interfaces/Organization/ProjectClaim/index.tsx
@@ -4,17 +4,24 @@ import { useMemo, useState } from 'react'
 import { ProjectClaimBenefits } from 'components/interfaces/Organization/ProjectClaim/benefits'
 import { ProjectClaimChooseOrg } from 'components/interfaces/Organization/ProjectClaim/choose-org'
 import { ProjectClaimConfirm } from 'components/interfaces/Organization/ProjectClaim/confirm'
-import { FormPanel } from 'components/ui/Forms/FormPanel'
 import ShimmeringLoader from 'components/ui/ShimmeringLoader'
-import { ApiAuthorizationResponse } from 'data/api-authorization/api-authorization-query'
+import { useApiAuthorizationQuery } from 'data/api-authorization/api-authorization-query'
 import { useOrganizationProjectClaimQuery } from 'data/organizations/organization-project-claim-query'
 import { useOrganizationsQuery } from 'data/organizations/organizations-query'
 import { Alert } from 'ui'
+import { ProjectClaimLayout } from './layout'
 
-export const ProjectClaim = ({ requester }: { requester: ApiAuthorizationResponse }) => {
-  const { token: claimToken } = useParams()
+export const ProjectClaim = () => {
+  const { auth_id, token: claimToken } = useParams()
   const [selectedOrgSlug, setSelectedOrgSlug] = useState<string>()
   const [step, setStep] = useState<'choose-org' | 'benefits' | 'confirm'>('choose-org')
+
+  const {
+    data: requester,
+    isLoading: isLoadingRequester,
+    isError: isErrorRequester,
+    error: errorRequester,
+  } = useApiAuthorizationQuery({ id: auth_id })
   const { data: organizations } = useOrganizationsQuery()
 
   const selectedOrganization = useMemo(() => {
@@ -24,8 +31,9 @@ export const ProjectClaim = ({ requester }: { requester: ApiAuthorizationRespons
   const {
     data: projectClaim,
     error: errorProjectClaim,
-    isError,
-    isLoading,
+    isError: isErrorProjectClaim,
+    isLoading: isLoadingProjectClaim,
+    isSuccess: isSuccessProjectClaim,
   } = useOrganizationProjectClaimQuery(
     {
       slug: selectedOrgSlug!,
@@ -35,6 +43,40 @@ export const ProjectClaim = ({ requester }: { requester: ApiAuthorizationRespons
       enabled: !!claimToken && !!selectedOrgSlug,
     }
   )
+
+  if ((selectedOrgSlug && claimToken && isLoadingProjectClaim) || isLoadingRequester) {
+    return (
+      <ProjectClaimLayout title="Claim a project">
+        <div className="py-6 space-y-2">
+          <ShimmeringLoader />
+          <ShimmeringLoader className="w-3/4" />
+          <ShimmeringLoader className="w-1/2" />
+        </div>
+      </ProjectClaimLayout>
+    )
+  }
+
+  if ((selectedOrgSlug && claimToken && isErrorProjectClaim) || isErrorRequester) {
+    return (
+      <ProjectClaimLayout title="Claim a project">
+        <div className="py-6">
+          <Alert
+            withIcon
+            variant="warning"
+            title="Failed to retrieve project claim request details"
+          >
+            <p>Please retry your claim request from the requesting app</p>
+            {errorProjectClaim != undefined && (
+              <p className="mt-2">Error: {errorProjectClaim?.message}</p>
+            )}
+            {errorRequester != undefined && (
+              <p className="mt-2">Error: {errorRequester?.message}</p>
+            )}
+          </Alert>
+        </div>
+      </ProjectClaimLayout>
+    )
+  }
 
   if (step === 'choose-org' || !selectedOrganization) {
     return (
@@ -47,53 +89,25 @@ export const ProjectClaim = ({ requester }: { requester: ApiAuthorizationRespons
     )
   }
 
-  if (isLoading) {
-    return (
-      <FormPanel header={<p>Claim a project</p>}>
-        <div className="px-8 py-6 space-y-2">
-          <ShimmeringLoader />
-          <ShimmeringLoader className="w-3/4" />
-          <ShimmeringLoader className="w-1/2" />
-        </div>
-      </FormPanel>
-    )
-  }
+  if (isSuccessProjectClaim) {
+    if (step === 'benefits') {
+      return (
+        <ProjectClaimBenefits
+          projectClaim={projectClaim}
+          requester={requester}
+          onContinue={() => setStep('confirm')}
+        />
+      )
+    }
 
-  if (isError) {
     return (
-      <FormPanel header={<p>Claim project</p>}>
-        <div className="px-8 py-6">
-          <Alert
-            withIcon
-            variant="warning"
-            title="Failed to retrieve project claim request details"
-          >
-            <p>Please retry your claim request from the requesting app</p>
-            {errorProjectClaim !== undefined && (
-              <p className="mt-2">Error: {errorProjectClaim?.message}</p>
-            )}
-          </Alert>
-        </div>
-      </FormPanel>
-    )
-  }
-
-  if (step === 'benefits') {
-    return (
-      <ProjectClaimBenefits
+      <ProjectClaimConfirm
+        setStep={setStep}
+        selectedOrganization={selectedOrganization}
         projectClaim={projectClaim}
         requester={requester}
-        onContinue={() => setStep('confirm')}
       />
     )
   }
-
-  return (
-    <ProjectClaimConfirm
-      setStep={setStep}
-      selectedOrganization={selectedOrganization}
-      projectClaim={projectClaim}
-      requester={requester}
-    />
-  )
+  return null
 }
