+++ apps/studio/components/layouts/APIAuthorizationLayout.tsx
@@ -18,20 +18,18 @@ const APIAuthorizationLayout = ({ children }: PropsWithChildren<APIAuthorization
         <div>
           <div className="mx-auto px-4 sm:px-6">
             <div className="max-w-xl flex justify-between items-center mx-auto py-4">
-              <div className="flex justify-start lg:w-0 lg:flex-1">
-                <div>
-                  <span className="sr-only">Supabase</span>
-                  <Image
-                    src={
-                      resolvedTheme?.includes('dark')
-                        ? `${BASE_PATH}/img/supabase-dark.svg`
-                        : `${BASE_PATH}/img/supabase-light.svg`
-                    }
-                    alt="Supabase Logo"
-                    height={20}
-                    width={105}
-                  />
-                </div>
+              <div className="flex justify-start lg:w-0 lg:flex-1 items-center">
+                <span className="sr-only">Supabase</span>
+                <Image
+                  src={
+                    resolvedTheme?.includes('dark')
+                      ? `${BASE_PATH}/img/supabase-dark.svg`
+                      : `${BASE_PATH}/img/supabase-light.svg`
+                  }
+                  alt="Supabase Logo"
+                  height={20}
+                  width={105}
+                />
               </div>
             </div>
           </div>
