+++ apps/studio/pages/project/[ref]/index.tsx
@@ -15,6 +15,7 @@ import { ComputeBadgeWrapper } from 'components/ui/ComputeBadgeWrapper'
 import { InlineLink } from 'components/ui/InlineLink'
 import { ProjectUpgradeFailedBanner } from 'components/ui/ProjectUpgradeFailedBanner'
 import { useEdgeFunctionsQuery } from 'data/edge-functions/edge-functions-query'
+import { useOAuthAppsQuery } from 'data/oauth/oauth-apps-query'
 import { createClaimToken } from 'data/projects/project-claim-token'
 import { useReadReplicasQuery } from 'data/read-replicas/replicas-query'
 import { useTablesQuery } from 'data/tables/tables-query'
@@ -79,25 +80,25 @@ const Home: NextPageWithLayout = () => {
   // [Joshen] JFYI minus 1 as the replicas endpoint returns the primary DB minimally
   const replicasCount = Math.max(0, (replicasData?.length ?? 1) - 1)
 
-  // const { data: apps } = useOAuthAppsQuery({ slug: organization?.slug })
+  const { data: apps } = useOAuthAppsQuery({ slug: organization?.slug })
 
   const buildOAuthUrl = async () => {
     try {
       const data = await createClaimToken({ projectRef: project?.ref! })
       console.log(data)
 
-      // const app = apps?.[0]
-      // if (!app) {
-      //   toast.error('No OAuth app found')
-      //   return
-      // }
-      // if (!app.redirect_uris) {
-      //   toast.error('No redirect URI found')
-      //   return
-      // }
+      const app = apps?.[0]
+      if (!app) {
+        toast.error('No OAuth app found')
+        return
+      }
+      if (!app.redirect_uris) {
+        toast.error('No redirect URI found')
+        return
+      }
 
-      // const url = `${process.env.NEXT_PUBLIC_API_ADMIN_URL}/oauth/authorize?client_id=5fce6198-92ec-443e-9b98-a5787827ac2a&response_type=code&redirect_uri=${app.redirect_uris[0]}`
-      // console.log(url)
+      const url = `${process.env.NEXT_PUBLIC_API_ADMIN_URL}/oauth/authorize?client_id=${app.client_id!}&response_type=code&redirect_uri=${app.redirect_uris[0]}`
+      console.log(url)
 
       // const response = await fetch(
       //   `${process.env.NEXT_PUBLIC_API_ADMIN_URL}/oauth/authorize?client_id=${app.app_id!}&response_type=code&redirect_uri=${app.redirect_uris[0]}`,
